backend/functions/pyproject.toml
[project]
name = "bacha-trainer-backend"
version = "1.0.0"
description = "Backend proxy for Bacha Trainer voice coaching with ElevenLabs and Google Gemini"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "elevenlabs>=1.0.0",
    "google-cloud-aiplatform>=1.38.0",
    "python-dotenv>=1.0.0",
    "mangum>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "hypothesis>=6.0.0",
    "httpx>=0.25.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
asyncio_mode = "auto"
